+++ include/LIEF/MachO.hpp
@@ -52,6 +52,7 @@
   #include "LIEF/MachO/FunctionVariantFixups.hpp"
   #include "LIEF/MachO/Header.hpp"
   #include "LIEF/MachO/IndirectBindingInfo.hpp"
+  #include "LIEF/MachO/LazyLoadDylibInfo.hpp"
   #include "LIEF/MachO/LinkEdit.hpp"
   #include "LIEF/MachO/LinkerOptHint.hpp"
   #include "LIEF/MachO/LoadCommand.hpp"